card flow wip
