+++ Podfile
@@ -4,7 +4,6 @@ platform :ios, '11.0'
 
 # Standard cocoapods specs source
 source 'https://github.com/CocoaPods/Specs.git'
- 
 
 use_frameworks!
 
@@ -20,8 +19,8 @@ def uisdkPods
   pod 'Adyen'
   pod 'PhoneNumberKit', '2.1.0'
  # pod 'KarhooSDK'
- pod 'KarhooSDK', :path => '../karhoo-ios-sdk'
-  # pod 'KarhooSDK', :git => 'git@github.com:karhoo/karhoo-ios-sdk.git', :branch => 'develop'
+ #pod 'KarhooSDK', :path => '../karhoo-ios-sdk'
+  pod 'KarhooSDK', :git => 'git@github.com:karhoo/karhoo-ios-sdk.git', :branch => 'develop'
 
 end
